move errors away from api folder & maintain only 1 file for exported errors
