--- ts/src/api/exceptions/errors/BadRequestError.ts
@@ -1,7 +1,0 @@
-import { CustomError, HttpCode } from "../CustomError";
-
-export class BadRequestError extends CustomError {
-  constructor(message: string){
-    super(message, HttpCode.BAD_REQUEST)
-  }
-}
