--- ts/src/api/exceptions/errors/NotFoundError.ts
@@ -1,7 +1,0 @@
-import { CustomError, HttpCode } from "../CustomError";
-
-export class NotFoundError extends CustomError {
-  constructor(message: string){
-    super(message, HttpCode.NOT_FOUND)
-  }
-}
