+++ ts/src/lib/errors/CustomError.ts
